Quiet Hands Demo
This is the quiet hands demo with the sound muting. It mutes the sound
of any other process other than the program itself. The way this was
done before was by process name so need to change that in the
documentation document. With the process name only was not possible to
mute everything but the program itself. Right now is using the process
ID.