Automated rebuilding of gesture interpretation files
Not yet finished. These two spikes work in conjunction with one another:
one forks the other which runs the KinectHandle and the other one plays
the clips. Right now you just select a .xef file and it plays it and
saves the gesture interpretation data in the corpus directory of
GestureTestingApp. I will continue working on this over the course of
the next few days to allow you to recursively run through all the clips
in a given directory and save the output to another given directory,
which can then be run through using the previous GestureTesting spike,
which is temporarily saved on my current machine. The importance of this
comes with training and rebuilding gesture databases. It will also be
possible to built only a subsection of the gestures or otherwise those
clips that satisfy certain properties.